backend/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-filter==21.1
djangorestframework==3.14.0
drf-base64==2.0
exceptiongroup==1.1.1
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
packaging==23.1
pep8-naming==0.13.3
pluggy==1.0.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytest==7.3.1
tomli==2.0.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7.1
reportlab==3.6.12
sqlparse==0.4.3
python-dotenv==0.20.0
djoser==2.1.0
fpdf==1.7.2
gunicorn==20.1.0